refactor: moved _extend
